requirements.txt
streamlit>=1.28.0
pdfplumber>=0.10.0
pandas>=2.0.0
openpyxl>=3.1.0
python-dotenv>=1.0.0
